Convert mouse navigation tests to Swift Testing
